backend/login_api/requirements.txt
fastapi
uvicorn
python-multipart

# extraction
PyMuPDF
python-docx
python-pptx
striprtf
beautifulsoup4
lxml
pandas
openpyxl
ebooklib
chardet

# OpenAI v1 SDK
openai>=1.0.0

